Enable split compilation in http::basic_fields
Moved functions,that are independent of template argument, into
an *.ipp file.
